+++ src/main/java/net/geforcemods/securitycraft/misc/StillValid.java
@@ -1,0 +1,7 @@
+package net.geforcemods.securitycraft.misc;
+
+import net.minecraft.world.entity.player.Player;
+
+public interface StillValid {
+	boolean stillValid(Player player);
+}